Renamed 'math' submodule to 'methods'
