+++ .build/build.spec
@@ -31,7 +31,7 @@ a = Analysis(
     pathex=[Path(build_dir), Path(project_root, "src")],
     binaries=[],
     datas=[],
-    hiddenimports=["fiat_build_time", "fiat.math"],
+    hiddenimports=["fiat_build_time", "fiat.methods"],
     hookspath=[build_dir.as_posix()],
     hooksconfig={},
     runtime_hooks=[Path(build_dir, 'runtime_hooks.py')],
