+++ src/fiat/methods/util.py
